Moved functions to transaction because the belong there
